feat: create code block display
