+++ package.json
@@ -48,6 +48,7 @@
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-helmet": "^6.1.0",
+    "react-syntax-highlighter": "^15.4.3",
     "tsconfig-paths-webpack-plugin": "3.3.0"
   },
   "devDependencies": {
@@ -57,6 +58,7 @@
     "@types/node": "14.14.31",
     "@types/react": "17.0.2",
     "@types/react-dom": "17.0.1",
+    "@types/react-syntax-highlighter": "^13.5.0",
     "@typescript-eslint/eslint-plugin": "4.16.1",
     "@typescript-eslint/parser": "4.16.1",
     "babel-plugin-root-import": "6.6.0",
